refactored package layout
