+++ lib/src/test/java/myJava/code/models/TestBinaryTreeLibrary.java
@@ -1,7 +1,7 @@
 /*
  * This Java source file was generated by the Gradle 'init' task.
  */
-package myJava.code.challenges;
+package myJava.code.models;
 
 import myJava.code.models.MyBinaryNode;
 import myJava.code.models.MyBinaryTree;
