+++ lib/src/test/java/myJava/code/models/TestKaryTreeLibrary.java
@@ -1,4 +1,4 @@
-package myJava.code.challenges;
+package myJava.code.models;
 
 import myJava.code.models.MyKaryNode;
 import org.junit.jupiter.api.Test;
