+++ lib/src/test/java/myJava/code/models/TestMyHashtable.java
@@ -1,4 +1,4 @@
-package myJava.code.challenges;
+package myJava.code.models;
 
 import myJava.code.models.MyHashtable;
 import org.junit.jupiter.api.Test;
